# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TNC Nursing Classes API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions:
    get:
      operationId: listSessions
      tags: [content]
      summary: Get sessions (videos/content) for a course
      parameters:
        - name: courseId
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
  /sessions/{rowId}:
    get:
      operationId: getSession
      tags: [content]
      summary: Get a single session by rowId
      parameters:
        - name: rowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /purchases/{userId}:
    get:
      operationId: getUserPurchases
      tags: [purchases]
      summary: Get purchased courses for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
  /purchases:
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Record a course purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Purchase recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "401":
          description: Invalid admin password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
...
  /promo/toggle:
    post:
      operationId: togglePromo
      tags: [promo]
      summary: Toggle promotional mode (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoToggleInput"
      responses:
        "200":
          description: Promo toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoStatus"
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [content]
      summary: List available quiz/exam sets
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Quiz list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizListResponse"
  /quiz/{examId}:
    get:
      operationId: getQuiz
      tags: [content]
      summary: Get a quiz/exam with all questions
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "404":
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: integer
        rowId:
          type: string
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: ["string", "null"]
        pdfUrl:
          type: ["string", "null"]
        contentType:
          type: string
          enum: ["youtube", "firebase", "pdf", "none"]
        type:
          type: string
        courseId:
          type: ["string", "null"]
        subjectId:
          type: ["string", "null"]
        isPaid:
          type: boolean
        duration:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        serialNo:
          type: string
        createdAt:
          type: string
      required:
        - id
        - rowId
        - title
        - isPaid
        - serialNo
        - contentType
...
    LoginInput:
      type: object
      properties:
        mobile:
          type: string
        password:
          type: string
      required:
        - mobile
        - password
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        password:
          type: string
        email:
          type: string
        college:
          type: string
        state:
          type: string
        city:
          type: string
        gender:
          type: string
        dob:
          type: string
      required:
        - name
        - mobile
        - password
    AuthResponse:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        token:
          type: string
      required:
        - userId
        - name
        - mobile
        - token
    Purchase:
      type: object
      properties:
        id:
          type: integer
        rowId:
          type: string
        userId:
          type: string
        courseId:
          type: string
        courseName:
          type: string
        amount:
          type: ["number", "null"]
        paymentId:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - rowId
        - userId
        - courseId
        - courseName
    PurchaseInput:
      type: object
      properties:
        userId:
          type: string
        courseId:
          type: string
        courseName:
          type: string
        amount:
          type: number
        paymentId:
          type: string
      required:
        - userId
        - courseId
        - courseName
    AdminLoginInput:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AdminAuthResponse:
      type: object
      properties:
        token:
          type: string
        message:
          type: string
      required:
        - token
        - message
...
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    PromoStatus:
      type: object
      properties:
        enabled:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        message:
          type: string
      required:
        - enabled
        - message
    PromoToggleInput:
      type: object
      properties:
        enabled:
          type: boolean
        durationDays:
          type: integer
        adminToken:
          type: string
      required:
        - enabled
        - adminToken
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    QuizListResponse:
      type: object
      properties:
        quizzes:
          type: array
          items:
            $ref: "#/components/schemas/QuizMeta"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - quizzes
        - total
        - page
        - limit
...
    Quiz:
      type: object
      properties:
        examId:
          type: string
        examNo:
          type: integer
        name:
          type: string
        maxMarks:
          type: number
        negativeMarks:
          type: number
        durationMinutes:
          type: string
        questionCount:
          type: integer
        validUntil:
          type: ["string", "null"]
        allowForPremium:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
      required:
        - examId
        - name
        - maxMarks
        - negativeMarks
        - durationMinutes
        - questions