# lib/api-zod/src/generated/types/adminAuthResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TNC Nursing Classes API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminAuthResponse {
  token: string;
  message: string;
}